requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.13.5
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
docx2pdf==0.1.8
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.37.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.71.2
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
lxml==6.0.1
mailmerge==2.2.4
Markdown==3.9
MarkupSafe==3.0.2
ordered-set==4.1.0
outcome==1.3.0.post0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyLaTeX==1.4.2
pyparsing==3.2.4
PySocks==1.7.1
python-docx==1.2.0
python-dotenv==1.1.1
pywin32==311
requests==2.32.5
rsa==4.9.1
selenium==4.35.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
tenacity==9.1.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing==3.7.4.3
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
